# personal-finance-management/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RESTful API Personal Finance Management",
    "description": "OpenAPI for Personal Finance Management, with the following features:\n1. User Management\n2. User Income\n3. User Expenses\n3. User Savings\n4. Personal Finance User Dashboard",
    "version": "1.0",
    "contact": {
      "name": "Mahfudz Nur Fadhila",
      "email": "mahfudnf63@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api"
    }
  ],
  "paths": {
...
    "/incomes" : {
      "post": {
        "tags": ["income"],
        "summary": "Create income",
        "description": "Create new income",
        "security": [
          {
            "bearerAuth" : []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json" : {
              "schema": {
                "type": "object",
                "required": ["amount","category"],
                "properties": {
                  "amount" : {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0,
                    "example": 100000
                  },
                  "category" : {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "gaji"
                  },
                  "description" : {
                    "type": "string",
                    "maxLength": 500,
                    "example": "pendapatan gaji bulan juni"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200" : {
            "description": "Create income success",
            "content": {
              "application/json" : {
                "schema": {
                  "$ref": "#/components/schemas/responseCreateIncomeSuccess"
                }
              }
            }
          },
          "401" : {
            "description": "Create income failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseUnauthorized"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["income"],
        "summary": "List incomes",
        "description": "Retrieve data according to the query parameters entered by the user",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "bearerAuth" : []
          }
        ],
        "responses": {
          "200" : {
            "description": "Success get list incomes",
            "content": {
              "application/json" : {
                "schema": {
                  "$ref": "#/components/schemas/responseListIncome"
                }
              }
            }
          },
          "401" : {
            "description": "Get list incomes failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseUnauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/incomes/{incomeId}" : {
      "get": {
        "tags": ["income"],
        "summary": "Get income",
        "description": "Retrieving income based on income id in path variable",
        "parameters": [
          {
            "name": "incomeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth" : []
          }
        ],
        "responses": {
          "200" : {
            "description": "Get income success",
            "content": {
              "application/json" : {
                "schema": {
                  "$ref": "#/components/schemas/responseCreateIncomeSuccess"
                }
              }
            }
          },
          "404" : {
            "description": "Get income not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseUnauthorized"
                },
                "example": {
                  "errors" : "Income not found"
                }
              }
            }
          },
          "401" : {
            "description": "Get income failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseUnauthorized"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["income"],
        "summary": "Edit income",
        "description": "Edit income based on income id in path variable",
        "parameters": [
          {
            "name": "incomeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth" : []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json" : {
              "schema": {
                "properties": {
                  "amount" : {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0,
                    "example": 200000
                  },
                  "category" : {
                    "type": "string",
                    "maxLength": 100,
                    "example": "freelance"
                  },
                  "description" : {
                    "type": "string",
                    "maxLength": 500,
                    "example": "gaji dari edit video"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200" : {
            "description": "Edit income success",
            "content": {
              "application/json" : {
                "schema": {
                  "$ref": "#/components/schemas/responseCreateIncomeSuccess"
                },
                "example": {
                  "data" : {
                    "id": "1231313",
                    "amount": 200000,
                    "category": "freelance",
                    "createdAt": "2026-02-23",
                    "description": "gaji dari edit video"
                  }
                }
              }
            }
          },
          "404" : {
            "description": "Get income not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseUnauthorized"
                },
                "example": {
                  "errors" : "Income not found"
                }
              }
            }
          },
          "401" : {
            "description": "Edit income failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseUnauthorized"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["income"],
        "summary": "Delete income",
        "description": "Delete income based on income id in path variable",
        "parameters": [
          {
            "name": "incomeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth" : []
          }
        ],
        "responses": {
          "200" : {
            "description": "Delete income success",
            "content": {
              "application/json" : {
                "schema": {
                  "$ref": "#/components/schemas/responseDataOk"
                }
              }
            }
          },
          "404" : {
            "description": "Get income not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseUnauthorized"
                },
                "example": {
                  "errors" : "Income not found"
                }
              }
            }
          },
          "401" : {
            "description": "Delete income failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseUnauthorized"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth" : {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "responseDataOk" : {
        "type": "object",
        "properties": {
          "data" : {
            "type": "string",
            "example": "OK"
          }
        }
      },
...
      "responseUnauthorized" : {
        "type": "object",
        "properties": {
          "errors": {
            "type": "string",
            "example": "Unauthorized"
          }
        }
      },
      "responseCreateIncomeSuccess" : {
        "type": "object",
        "properties": {
          "data" : {
            "type": "object",
            "properties": {
              "incomeId" : {
                "type": "string",
                "example": "1231313"
              },
              "amount" : {
                "type": "integer",
                "format": "int64",
                "example": 100000
              },
              "category" : {
                "type": "string",
                "example": "gaji"
              },
              "createdAt" : {
                "type": "string",
                "format": "date",
                "example": "2026-02-23"
              },
              "description" : {
                "type": "string",
                "example": "pendapatan gaji bulan juni"
              }
            }
          }
        }
      },
      "paging" : {
        "type": "object",
        "properties": {
          "currentPage" : {
            "type": "integer",
            "example": 0
          },
          "totalPages" : {
            "type": "integer",
            "example": 10
          },
          "size" : {
            "type": "integer",
            "example": 10
          }
        }
      },
      "responseListIncome" : {
        "type": "object",
        "properties": {
          "data" : {
            "type": "array",
            "items": {
              "properties": {
                "incomeId" : {
                  "type": "string",
                  "example": "1231313"
                },
                "amount" : {
                  "type": "integer",
                  "format": "int64",
                  "example": 100000
                },
                "category" : {
                  "type": "string",
                  "example": "gaji"
                },
                "createdAt" : {
                  "type": "string",
                  "format": "date",
                  "example": "2026-02-23"
                },
                "description" : {
                  "type": "string",
                  "example": "pendapatan gaji bulan juni"
                }
              }
            }
          },
          "paging" : {
            "$ref": "#/components/schemas/paging"
          }
        }
      },
...
    }
  }
}